support writing image tilesets as JPEG
git-svn-id: http://vtp.googlecode.com/svn/trunk@6881 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/ImageLayer.cpp
@@ -163,7 +163,7 @@ void vtImageLayer::GetPropertyText(wxString &strIn)
 
 	strIn += _("Bitmaps:");
 	strIn += _T("\n");
-	for (int i = 0; i < m_pImage->NumBitmaps(); i++)
+	for (size_t i = 0; i < m_pImage->NumBitmaps(); i++)
 	{
 		BitmapInfo &bmi = m_pImage->GetBitmapInfo(i);
